+++ src/finlang/Transpiler/TranspilerOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace finlang.Transpiler;
+
+public class TranspilerOptions
+{
+    public bool OutputVersionInfo = true;
+    public bool OutputTimestamp = true;
+    public bool OutputChecksum = true;
+}